Extract diff line stats helper
